size and style
