+++ app/design-system/design-system-hedvig/src/main/kotlin/com/hedvig/android/design/system/hedvig/tokens/AnimationTokens.kt
@@ -1,5 +1,8 @@
 package com.hedvig.android.design.system.hedvig.tokens
 
+import androidx.compose.animation.core.tween
+
 internal class AnimationTokens {
   val fastAnimationDuration: Int = 200
+  val pulsatingAnimationDuration: Int = 400
 }
